requirements.txt
kagglehub>=0.1.0
shapely>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
scipy>=1.9.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.64.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.15.0
